move clampheading to common and fix simulator bot rotation
