+++ src/common/game-types.ts
@@ -1,6 +1,6 @@
 import { Square } from "chess.js";
 import type { Robot } from "../server/robot/robot";
-import { DEGREE } from "../server/utils/units";
+import { DEGREE } from "./units";
 
 /**
  * Defines a specific piece.
